pyln: Add a version parser to pyln-testing
We changed the way we configure CLN to run in developer mode, which
promptly broke anything that wasn't v23.11 and later. So we parse the
version, and make it comparable, so we can determine what to do on the
fly, and maintain a better backwards compatibility.

Changelog-Changed: pyln-testing: pyln-testing is now version-aware and can customize the run based on the version.